+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ### v1.2 (2014-03-11) (Dev time: 5 hours)
 - Add: Add the ability to specify the PHP script timeout. Thanks https://github.com/grantnorwood
+- Add: Add uninstall.php file to remove option when plugin is removed. Thanks https://github.com/grantnorwood
 
 ### v1.1.1 (2014-02-12) (Dev time: 0.5 hours)
 - Add: Add German translation